Fix: change http status 200 -> 201 Created
