+++ controllers/animal.ctrl.js
@@ -44,6 +44,52 @@ exports.getUserAnimal = async function (req, res, next) {
 	}
 };
 
+exports.mergeAnimal = async function (req, res, next) {
+	const { animalId1, animalId2, tokenURL } = req.body;
+	if (animalId1 === undefined || animalId2 === undefined) {
+		return res.status(400).send(errorMsg.needParameter);
+	}
+	try {
+		// NFT가 아니고, 두가지 속성중 무작위 하나 선택
+		let animals = await models.animal_possession.findAll({
+			where: { [Op.or]: [{ id: animalId1 }, { id: animalId2 }] },
+		});
+		let animal_type = animals[randomVal()].animal_type;
+		let color = await colorUtils.synthesizeColor(animals[0].dataValues.color, animals[1].dataValues.color, animal_type);
+		let user_wallet = await models.user.findOne({ where: { id: req.userId } });
+
+		function randomVal() {
+			return Math.round(Math.random());
+		}
+		let new_animal = {
+			name: animals[randomVal()].name,
+			tier: animals[randomVal()].tier,
+			user_id: req.userId,
+			color: color,
+			animal_type: animal_type,
+			head_item_id: animals[randomVal()].head_item_id,
+			pattern_id: animals[randomVal()].pattern_id,
+		};
+
+		let nftMintResult = await nftUtils.getNft(title = 'Bluehat Animal', symbol = 'Bluehat', tokenURL, toAddr = user_wallet.wallet_address);
+		new_animal['nft_hash'] = nftMintResult.transactionHash;
+
+		let new_animals = await models.animal_possession.create(new_animal);
+
+		await models.animal_possession
+			.destroy({
+				where: { [Op.or]: [{ id: animalId1 }, { id: animalId2 }] },
+			})
+			.then(logger.info("merge success"));
+
+		return res.status(201).send(new_animals);
+
+	} catch (e) {
+		logger.error(e);
+		return res.status(500).send(errorMsg.internalServerError);
+	}
+};
+
 exports.makeNewAnimal = async function (req, res, next) {
 	/*
 		1. Create New Animal to user
@@ -92,9 +138,8 @@ exports.changeAnimalColor = async function (req, res, next) {
 			attributes: ["animal_type", "color"],
 		});
 		let new_color = await colorUtils.changeColor(animal.dataValues.color, animal.dataValues.animal_type);
-		console.log("NEW_COLOR: ", new_color);
 		await models.animal_possession.update({ color: new_color }, { where: { id: animal_id } });
-		return res.status(200).send(infoMsg.success);
+		return res.status(201).send(infoMsg.success);
 	} catch (e) {
 		logger.error(`${req.method} ${req.url}` + ": " + e);
 		return res.status(500).send(errorMsg.internalServerError);
